third commit

+++ controller/contactController.js
@@ -1,5 +1,6 @@
 const asyncHandler =require('express-async-handler');
-const Contact = require("../models/contactModels")
+const Contact = require("../models/contactModel");
+const { default: mongoose } = require('mongoose');
 //@desc create contact
 //@rout POST /api/contact
 //@access public
@@ -12,7 +13,7 @@ const createContact =asyncHandler(async(req,res)=>{
     }
    else
    {
-        const contact = await Contact.create({name,email,phone});
+        const contact = await Contact.create({name,email,phone,user_id:req.user.id});
         console.log("created.");
         res.status(201).json(contact);
    }
@@ -23,7 +24,8 @@ const createContact =asyncHandler(async(req,res)=>{
 //@rout GET /api/contact/:id
 //@access public
 const getContact =asyncHandler(async(req,res)=>{
-    const contact = await Contact.findById(req.params.id);
+    const objectid =new mongoose.Types.ObjectId(req.params.id);
+    const contact = await Contact.find(objectid);
     if(!contact)
     {
         res.status(404);
@@ -36,7 +38,7 @@ const getContact =asyncHandler(async(req,res)=>{
 //@rout GET /api/contacts
 //@access public
 const getContacts =asyncHandler(async(req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id:req.user.id});
     res.status(200).json({contacts});
 })
 
@@ -53,15 +55,20 @@ const updateContact =asyncHandler(async(req,res)=>{
         res.status(404);
         throw new Error("Contact Not Found.");   
      }
-    else if(!name || !email || !phone)
+
+    if(!name || !email || !phone)
     {
         res.status(400);
         throw new Error("All Fields Are Mandatory !");
     }
-    else{
+     if(contact.user_id.toString()!==req.user.id)
+     {
+        res.status(403);
+        throw new Error("User Dont have Permission to modify this contact");
+     }
         const updatedContact = await Contact.findByIdAndUpdate(id,req.body,{new:true});
         res.status(200).json(updatedContact);
-    }
+    
 })
 
 //@desc delete  contact
@@ -76,11 +83,15 @@ const deleteContact =asyncHandler(async(req,res)=>{
         res.status(404);
         throw new Error("Contact Not Found.");   
      }
-     else
-     {
+    
+     if(contact.user_id.toString()!==req.user.id)
+        {
+           res.status(403);
+           throw new Error("User Dont have Permission to modify this contact");
+        }
         const deletedContact = await Contact.findByIdAndDelete(id,contact);
         res.status(200).json(deletedContact);   
-     }
+     
 })
 
 module.exports={createContact,getContact,getContacts,updateContact,deleteContact};
